requirements.txt
pandas
numpy
streamlit
supabase
python-telegram-bot
requests
python-dotenv
plotly
matplotlib
